Replace long int with long and off_t types
Replace "long int" with "long" and use proper POSIX types where
appropriate to improve code consistency and type correctness

Changes:
- journal_utils.hpp: Changed "long int" to "long" for severity
  variable and getJournalMetadataInt() parameter
- log_services_utils.hpp: Changed "long long int" to "off_t" for
  lseek() return values to match POSIX standard type
- ssl_key_handler.cpp: Changed "long int" to "long" for
  BIO_get_mem_data() return value

These changes improve:
1. Type safety by using off_t for file offsets
2. Code readability by removing redundant "int" keyword
3. Consistency with existing code (line 55 in log_services_utils.hpp
   already used off_t correctly)
4. Standards compliance with C++ best practices

Tested: Built successfully, verified all function signatures match
their respective API expectations

+++ redfish-core/include/utils/log_services_utils.hpp
@@ -34,7 +34,7 @@ namespace log_services_utils
 
 inline bool checkSizeLimit(int fd, crow::Response& res)
 {
-    long long int size = lseek(fd, 0, SEEK_END);
+    off_t size = lseek(fd, 0, SEEK_END);
     if (size <= 0)
     {
         BMCWEB_LOG_ERROR("Failed to get size of file, lseek() returned {}",
@@ -44,7 +44,7 @@ inline bool checkSizeLimit(int fd, crow::Response& res)
     }
 
     // Arbitrary max size of 20MB to accommodate BMC dumps
-    constexpr long long int maxFileSize = 20LL * 1024LL * 1024LL;
+    constexpr off_t maxFileSize = 20LL * 1024LL * 1024LL;
     if (size > maxFileSize)
     {
         BMCWEB_LOG_ERROR("File size {} exceeds maximum allowed size of {}",
